checking of home page
