+++ lib/post-utils.js
@@ -64,6 +64,16 @@ export function getTopicPosts(topic) {
     return topicPost;
 }
 
+export function getAllTopics() {
+    var tab = []
+    const allPosts = getAllPosts();
+    
+    tab.push(allPosts.map((post => post.topics)))
+
+    return tab;
+
+}
+
 export function getTags() {
     
 }
