+++ pages/[postId].js
@@ -1,8 +1,8 @@
-import { getPostData, getPostFiles } from '../../lib/post-utils';
-import PostContent from '../../components/post-detail/post-content';
-import PostTags from '../../components/post-detail/post-tags';
+import { getPostData, getPostFiles } from '../lib/post-utils';
+import PostContent from '../components/post-detail/post-content';
+import PostTags from '../components/post-detail/post-tags';
 import { Fragment } from 'react';
-import Comments from '../../components/input/comments';
+import Comments from '../components/input/comments';
 
 import Head from 'next/head';
